Extract generatepdf functionality from data service
