+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppssubjectaccessrequestworker/services/GetSubjectAccessRequestDataService.kt
@@ -1,15 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppssubjectaccessrequestworker.services
 
-import com.itextpdf.text.BaseColor
-import com.itextpdf.text.Chunk
-import com.itextpdf.text.Document
-import com.itextpdf.text.Font
-import com.itextpdf.text.FontFactory
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppssubjectaccessrequestworker.gateways.GenericHmppsApiGateway
-import java.io.ByteArrayOutputStream
 import java.time.LocalDate
 
 @Service
@@ -30,29 +24,5 @@ class GetSubjectAccessRequestDataService(@Autowired val genericHmppsApiGateway: 
       response?.get("content")?.let { responseObject[service] = it }
     }
     return responseObject
-  }
-  fun generatePDF(
-    content: Map<String, Any>,
-    document: Document = Document(),
-    pdfStream: ByteArrayOutputStream = ByteArrayOutputStream(),
-    pdfService: PdfService = PdfService(),
-  ): ByteArrayOutputStream {
-    log.info("Saving report..")
-    pdfService.getPdfWriter(document, pdfStream)
-    document.open()
-    log.info("Started writing to PDF")
-    val font: Font = FontFactory.getFont(FontFactory.COURIER, 16f, BaseColor.BLACK)
-    log.info("Set font")
-    if (content == emptyMap<Any, Any>()) {
-      document.add(Chunk("NO DATA FOUND", font))
-    }
-    content.forEach { entry ->
-      log.info(entry.key + entry.value)
-      document.add(Chunk("${entry.key} : ${entry.value}", font))
-    }
-    log.info("Finished writing report")
-    document.close()
-    log.info("PDF complete")
-    return pdfStream
   }
 }
